Рефакторинг. Никакой новой функциональности добавлено не было. Была реорганизована структура хранения. command/handler - обработчики команд, command/parser - парсеры команд.
Теперь main описывает обработчики команд как дерево с ветками, где каждой команде соотв. свой обработчик.

Удалил pretty_env_logger.
