+++ src/command/handler/mod.rs
@@ -20,8 +20,14 @@
  *
  *
  */
-
-pub mod command;
-pub mod error;
-pub mod listener;
-pub mod model;
+pub mod decode_handler;
+pub mod encode_handler;
+pub mod help_handler;
+pub mod radix_handler;
+pub mod switch_keyboard_handler;
+pub mod tracking_handler;
+pub mod unixtime_handler;
+pub mod winner_handler;
+pub mod qr_handler;
+pub mod jp_handler;
+pub mod transform_handler;
