+++ src/command/handler/transform_handler.rs
@@ -21,34 +21,13 @@
  *
  */
 
-use crate::command::BezzabotError;
+use crate::error::BezzabotError;
 use crate::model::markers::{DavinciMarker, YoutubeTimeTag};
 use csv::ReaderBuilder;
-use teloxide::utils::command::ParseError;
-
-type TransformFrom = String;
-type TransformTo = String;
-
-const INCORRECT_FORMAT_MESSAGE: &str =
-    r#"Неверный формат. Поддерживаются преобразования davinci -> yt"#;
-
-pub fn transform_parser(input: String) -> Result<(TransformFrom, TransformTo), ParseError> {
-    let args: Vec<&str> = input.split_whitespace().collect();
-
-    if args.len() != 2 {
-        return Err(ParseError::IncorrectFormat(INCORRECT_FORMAT_MESSAGE.into()));
-    }
-
-    let (from, to) = (args[0], args[1]);
-
-    if from != "davinci" && to != "yt" {
-        return Err(ParseError::IncorrectFormat(
-            "Доступные форматы davinci -> yt".into(),
-        ));
-    }
-
-    Ok((from.into(), to.into()))
-}
+use teloxide::net::Download;
+use teloxide::prelude::{Message, Requester, ResponseResult};
+use teloxide::types::{MediaKind, MessageKind};
+use teloxide::{respond, Bot};
 
 pub struct DavinciYoutubeTransformer;
 
@@ -57,8 +36,8 @@ impl<'a> DavinciYoutubeTransformer {
         let mut reader = ReaderBuilder::new().from_reader(bytes);
         let mut lines = vec![];
 
-        for davinci in reader.deserialize::<DavinciMarker>() {
-            let marker = davinci.map_err(BezzabotError::CsvError)?;
+        for davinci_result in reader.deserialize::<DavinciMarker>() {
+            let marker = davinci_result.map_err(BezzabotError::CsvError)?;
             lines.push(YoutubeTimeTag::from_davinci(marker))
         }
 
@@ -69,4 +48,34 @@ impl<'a> DavinciYoutubeTransformer {
             .join("\n");
         Ok(result)
     }
+}
+
+pub async fn transform_handler(bot: Bot, msg: Message) -> ResponseResult<()> {
+    let MessageKind::Common(mc) = msg.kind else {
+        return respond(());
+    };
+
+    let MediaKind::Document(md) = mc.media_kind else {
+        return respond(());
+    };
+
+    let caption = md.caption.unwrap_or(String::from("davinci -> yt"));
+    let (from, to) = caption.split_once(',').unwrap_or(("davinci", "yt"));
+
+    let mut dst: Vec<u8> = vec![];
+
+    let transformer = DavinciYoutubeTransformer;
+
+    if from == "davinci" && to == "yt" {
+        let file = bot.get_file(md.document.file.id).await?;
+        bot.download_file(&file.path, &mut dst).await?;
+    }
+
+    let result = transformer.transform(&dst).await;
+
+    let transformed = result.unwrap_or_else(|e| e.to_string());
+
+    bot.send_message(msg.chat.id, transformed).await?;
+
+    respond(())
 }
