+++ src/command/parser/mod.rs
@@ -20,8 +20,9 @@
  *
  *
  */
-
-pub mod command;
-pub mod error;
-pub mod listener;
-pub mod model;
+pub mod encdec_parser;
+pub mod radix_parser;
+pub mod switch_keyboard_parser;
+pub mod tracking_parser;
+pub mod transform_parser;
+pub mod winner_parser;
