+++ src/error.rs
@@ -1,0 +1,65 @@
+/*
+ * ______          _ _
+ * | ___ \        | (_)
+ * | |_/ /__ _  __| |_  ___  _ __   __ _ _ __  _   _ ___
+ * |    // _` |/ _` | |/ _ \| '_ \ / _` | '_ \| | | / __|
+ * | |\ \ (_| | (_| | | (_) | |_) | (_| | |_) | |_| \__ \
+ * \_| \_\__,_|\__,_|_|\___/| .__/ \__,_| .__/ \__,_|___/
+ *                          | |         | |
+ *                          |_|         |_|
+ *
+ * twitch: twitch.tv/radiopapus
+ * github: https://github.com/radiopapus
+ * telegram: https://t.me/radiopapus
+ *
+ * Отказ от ответственности - Использовать только в образовательных целях. Распространяется "как есть".
+ *
+ * Disclaimer - Only for educational purposes.
+ *
+ * 2023.
+ *
+ *
+ */
+
+/*
+ * ______          _ _
+ * | ___ \        | (_)
+ * | |_/ /__ _  __| |_  ___  _ __   __ _ _ __  _   _ ___
+ * |    // _` |/ _` | |/ _ \| '_ \ / _` | '_ \| | | / __|
+ * | |\ \ (_| | (_| | | (_) | |_) | (_| | |_) | |_| \__ \
+ * \_| \_\__,_|\__,_|_|\___/| .__/ \__,_| .__/ \__,_|___/
+ *                          | |         | |
+ *                          |_|         |_|
+ *
+ * twitch: twitch.tv/radiopapus
+ * github: https://github.com/radiopapus
+ * telegram: https://t.me/radiopapus
+ *
+ * Отказ от ответственности - Использовать только в образовательных целях. Распространяется "как есть".
+ *
+ * Disclaimer - Only for educational purposes.
+ *
+ * 2023.
+ *
+ *
+ */
+
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum BezzabotError {
+    #[error("fetch data error ")]
+    FetchError(#[from] reqwest::Error),
+
+    #[error("serde data error ")]
+    SerdeError(#[from] serde_json::Error),
+
+    #[error("csv error ")]
+    CsvError(#[from] csv::Error),
+
+    #[error("parse data error `{0}`")]
+    ParseError(String),
+    
+    #[error("unknown error")]
+    Unknown,
+}
